rough sketch of the idea
A simple procedural implementation to test the general idea is a great
way to start a project. Worrying about the "architecture" too early in
the project lifecycle wastes time and is likely to lead down the wrong
path. It's hard to know in advance which pieces of code to generalize.
It's easy to accidentally make your code overly-abstract when imagining
the project while staring at a blank document.